+++ .gitignore
@@ -1,3 +1,4 @@
+lib/
 # Logs
 logs
 *.log
