--- iouring-engineering__ui-components.d.ts
@@ -1,1 +1,0 @@
-declare module '@iouring-engineering/ui-components';
