+++ package.json
@@ -10,31 +10,36 @@
     "market heatmap"
   ],
   "scripts": {
-    "build": "rollup -c"
+    "build": "rimraf lib && rollup -c"
   },
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
+  "main": "lib/index.js",
+  "module": "lib/index.esm.js",
+  "types": "lib/index.d.ts",
   "typings": "iouring-engineering__ui-components.d.ts",
   "author": "Devika Das",
   "license": "ISC",
   "dependencies": {
     "@mui/material": ">=5.13.4",
+    "rimraf": "^5.0.5"
+  },
+  "peerDependencies": {
+    "@types/react": ">=17.0.43",
     "react": ">=17.0.2",
     "react-dom": ">=17.0.2"
   },
   "devDependencies": {
-    "postcss-scss": ">=4.0.3",
-    "rollup": "^2.79.1",
     "@rollup/plugin-commonjs": "^25.0.7",
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@rollup/plugin-typescript": "^11.1.5",
+    "@types/react-dom": ">=17.0.14",
+    "postcss-scss": ">=4.0.3",
+    "rollup": "^2.79.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-preserve-directives": "^0.2.0",
     "rollup-plugin-terser": "^7.0.2",
     "sass": ">=1.49.9",
     "tslib": "^2.6.2",
-    "@types/react": ">=17.0.43",
-    "@types/react-dom": ">=17.0.14",
     "typescript": ">=4.7.4"
   }
 }
