+++ src/heatmap/index.tsx
@@ -1,0 +1,1 @@
+export { HeatmapContentNodeProps, Heatmap, HeatmapFilter } from "./Heatmap"
