+++ tsconfig.json
@@ -1,19 +1,15 @@
 {
   "compilerOptions": {
-    // Default
     "target": "es5",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true,
-
-    // Added
     "jsx": "react",
     "module": "ESNext",
     "declaration": true,
-    "declarationDir": "types",
     "sourceMap": false,
-    "outDir": "dist",
+    "outDir": "lib",
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
     "emitDeclarationOnly": true,
